requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
requests==2.32.3
celery==5.4.0
redis==5.0.7
django-redis==5.4.0
whitenoise==6.7.0
gunicorn==22.0.0
drf-yasg==1.21.5
django-filter==2.4.0
django-celery-beat==2.8.0

python-dotenv==1.0.1

setuptools>=70.0.0

wheel
